Update the trusted certs recipe to copy in certs that were linked in from external folders

Fixes https://gitlab.com/gitlab-org/omnibus-gitlab/issues/1379 and  Fixes https://gitlab.com/gitlab-org/omnibus-gitlab/issues/1378



See merge request !877

+++ files/gitlab-cookbooks/gitlab/libraries/helper.rb
@@ -458,7 +458,7 @@ class CertificateHelper
   def move_existing_certificates
     Dir.glob(File.join(@omnibus_certs_dir, "*")) do |file|
       case
-      when !valid?(file),whitelisted_files.include?(file)
+      when !valid?(file),whitelisted_files.include?(File.realpath(file))
         next
       when is_x509_certificate?(file)
         move_certificate(file)
@@ -469,18 +469,26 @@ class CertificateHelper
   end
 
   def valid?(file)
-    if File.symlink?(file) && !File.exists?(file)
-      FileUtils.rm_f(file)
-      return false
-    end
+    exists = File.exists?(file)
+    FileUtils.rm_f(file) if File.symlink?(file) && !exists
 
-    true
+    exists
   end
 
   def move_certificate(file)
     return if File.symlink?(file) && File.readlink(file).start_with?(@trusted_certs_dir)
-    FileUtils.mv(file, @trusted_certs_dir, force: true)
-    puts "\n Moving #{file}"
+
+    # Move the certs to the trusted certs directory if it is located within our managed certs directory
+    # Otherwise copy the cert to the trusted certs directory
+    realpath = File.realpath(file)
+    if realpath.start_with?(@omnibus_certs_dir)
+      FileUtils.mv(realpath, @trusted_certs_dir, force: true)
+    else
+      FileUtils.cp(realpath, @trusted_certs_dir)
+    end
+
+    FileUtils.rm_f(file) if File.symlink?(file)
+    puts "\n Moving #{realpath}"
   end
 
   def link_certificates
